Added WeatherListDTO Tests as well as modifications to add field missing in WeatherListDTO and double used for selecting multiple cities by coordinates in a circle

+++ src/test/java/com/spartaglobal/eng76/framework/urlbuilder/URLBuilder.java
@@ -174,7 +174,7 @@ public class URLBuilder {
         return create(BASE_URL + "box/city?" + "bbox=" + lon_left + "," + lat_bottom + "," + lon_right + "," + lat_top + "," + zoom, apikey);
     }
 
-    public static URLBuilder ofCitiesInCircle(long lat, long lon, String apikey) {
+    public static URLBuilder ofCitiesInCircle(double lat, double lon, String apikey) {
         return create(BASE_URL + "find?" + "lat=" + lat + "&lon=" + lon, apikey);
     }
